add is_hot

+++ src/main/java/com/example/backend/service/AttributeValueService.java
@@ -6,6 +6,7 @@ import com.example.backend.entity.Attribute;
 import com.example.backend.entity.AttributeValue;
 import com.example.backend.repository.AttributeRepo;
 import com.example.backend.repository.AttributeValueRepo;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -50,10 +51,21 @@ public class AttributeValueService {
                 attribute.getId(), updatedAttributeValue.getValue());
     }
 
-    public void deleteAttributeValue(String id) {
-        AttributeValue attributeValue = attributeValueRepo.findById(id)
-                .orElseThrow(() -> new RuntimeException("AttributeValue not found"));
-        attributeValueRepo.delete(attributeValue);
+    public void deleteAttributeValue(String attributeValueId) {
+        AttributeValue attributeValue = attributeValueRepo.findById(attributeValueId)
+                .orElseThrow(() -> new EntityNotFoundException("Brand not found with id: " + attributeValueId));
+
+        // Kiểm tra xem brand có liên kết với sản phẩm nào không
+        boolean hasAttributeValues = attributeValueRepo.existsProductsByAttributeValueId(attributeValueId);
+
+        if (hasAttributeValues) {
+            // Cập nhật trường is_deleted nếu có liên kết
+            attributeValue.set_deleted(true);
+            attributeValueRepo.save(attributeValue);
+        } else {
+            // Xóa hoàn toàn nếu không có liên kết
+            attributeValueRepo.delete(attributeValue);
+        }
     }
 
     public AttributeValueRes getAttributeValueById(String id) {
@@ -64,7 +76,7 @@ public class AttributeValueService {
     }
 
     public List<AttributeValueRes> getAllAttributeValues() {
-        List<AttributeValue> attributeValues = attributeValueRepo.findAll();
+        List<AttributeValue> attributeValues = attributeValueRepo.findAllActiveAttributeValues();
         return attributeValues.stream()
                 .map(attributeValue -> new AttributeValueRes(attributeValue.getId(),
                         attributeValue.getAttribute().getId(), attributeValue.getValue()))
